Make diagnostic code cross-platform.
