+++ Library/Homebrew/extend/os/diagnostic.rb
@@ -1,0 +1,5 @@
+require "diagnostic"
+
+if OS.mac?
+  require "extend/os/mac/diagnostic"
+end
